+++ gui/src/message/message_object.rs
