+++ gui/src/user/mod.rs
@@ -1,9 +1,9 @@
-pub mod user_data;
+pub mod user_object;
 pub mod user_profile;
 pub mod user_prompt;
 pub mod user_row;
 
-pub use user_data::UserObject;
+pub use user_object::UserObject;
 pub use user_profile::UserProfile;
 pub use user_prompt::UserPrompt;
 pub use user_row::UserRow;
